requirements.txt
Flask==1.0.3
Flask-Mail==0.9.1
PyGithub==1.43.7
PyYAML==5.1.1
